refactor(elb): move data_source elb flavors to service path
