+++ huaweicloud/provider.go
@@ -389,7 +389,7 @@ func Provider() *schema.Provider {
 			"huaweicloud_dms_product":         dms.DataSourceDmsProduct(),
 			"huaweicloud_dms_maintainwindow":  dms.DataSourceDmsMaintainWindow(),
 
-			"huaweicloud_elb_flavors":        dataSourceElbFlavorsV3(),
+			"huaweicloud_elb_flavors":        elb.DataSourceElbFlavorsV3(),
 			"huaweicloud_elb_pools":          elb.DataSourcePools(),
 			"huaweicloud_enterprise_project": eps.DataSourceEnterpriseProject(),
 			"huaweicloud_evs_volumes":        evs.DataSourceEvsVolumesV2(),
